Added status indicator and message to status tab. Fixed bug preventing page load for non-logged-in users
