+++ web/frontend/src/utils.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import { message } from 'antd';
-export const displayError = (error: Error) => message.error(error.toString());
+
+export const displayError = (error: Error) => message.error(error.toString(), 5);
 
 export const jwtFetch = (path: string, options?: object) => {
   const jwt = window.localStorage.getItem('jwt');
   if (jwt === null) {
     const error = 'JWT not found in local storage. You must be logged in.';
     message.error(error);
-    throw error;
   }
   const newOptions = {
     ...options,
